unit_tests int64

+++ src/smoke_detection/pipelines/data_unit_tests/nodes.py
@@ -20,21 +20,21 @@ def unit_test(
     assert pd_df_ge.expect_table_column_count_to_equal(16).success == True
     
     # Assert data types
-    assert pd_df_ge.expect_column_values_to_be_of_type("UTC", "int").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("UTC", "int64").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("Temperature[C]", "float").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("Humidity[%]", "float").success == True
-    assert pd_df_ge.expect_column_values_to_be_of_type("TVOC[ppb]", "int").success == True
-    assert pd_df_ge.expect_column_values_to_be_of_type("eCO2[ppm]", "int").success == True
-    assert pd_df_ge.expect_column_values_to_be_of_type("Raw H2", "int").success == True
-    assert pd_df_ge.expect_column_values_to_be_of_type("Raw Ethanol", "int").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("TVOC[ppb]", "int64").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("eCO2[ppm]", "int64").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("Raw H2", "int64").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("Raw Ethanol", "int64").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("Pressure[hPa]", "float").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("PM1.0", "float").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("PM2.5", "float").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("NC0.5", "float").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("NC1.0", "float").success == True
     assert pd_df_ge.expect_column_values_to_be_of_type("NC2.5", "float").success == True
-    assert pd_df_ge.expect_column_values_to_be_of_type("CNT", "int").success == True
-    assert pd_df_ge.expect_column_values_to_be_of_type("Fire Alarm", "int").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("CNT", "int64").success == True
+    assert pd_df_ge.expect_column_values_to_be_of_type("Fire Alarm", "int64").success == True
     
     # Assert column uniqueness to detect duplicates
     assert pd_df_ge.expect_column_values_to_be_unique("UTC").success == True
